chore: moved redis to config
