+++ src/jobs/queues/postView.queue.ts
@@ -1,4 +1,4 @@
-import redisConnection from '#src/redis.js';
+import redisConnection from '#src/config/redis.js';
 import { Queue } from 'bullmq';
 
 export const postViewCleanupQueue = new Queue('cleanup-postviews', { connection: redisConnection });
