+++ src/jobs/queues/test.queue.ts
@@ -1,4 +1,4 @@
-import redisConnection from '#src/redis.js';
+import redisConnection from '#src/config/redis.js';
 import { Queue } from 'bullmq';
 
 export const testQueue = new Queue('test-queue', { connection: redisConnection });
