+++ src/jobs/queues/updateTaggedPosts.queue.ts
@@ -1,5 +1,5 @@
 import { Queue } from 'bullmq';
-import redisConnection from '#src/redis.js';
+import redisConnection from '#src/config/redis.js';
 
 export const updateTaggedPostsQueue = new Queue('refreshPosts', {
   connection: redisConnection,
