+++ src/jobs/workers/postPopularity.worker.ts
@@ -1,5 +1,5 @@
 import postPopularityStrategy from '#src/recommendations/strategies/postPopularity.strategy.js';
-import redisConnection from '#src/redis.js';
+import redisConnection from '#src/config/redis.js';
 import { Worker } from 'bullmq';
 
 export const cleanupWorker = new Worker(
