+++ src/recommendations/strategies/home.strategy.ts
@@ -3,7 +3,7 @@ import { CommunityFollowedBy } from '#src/models/communityFollowedBy.model.js';
 import Post from '#src/models/post.model.js';
 import { PostView } from '#src/models/postView.model.js';
 import { UserFollows } from '#src/models/userFollows.model.js';
-import redisConnection from '#src/redis.js';
+import redisConnection from '#src/config/redis.js';
 
 export const computeHomeFeed = async (userId: string) => {
   console.log('computing home feed for user', userId);
